feat: 新增扇形火焰、火焰雷射和著火效果

+++ Assets/Scripts/Skill/Addons/AcidBombAddon.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 public class AcidBombAddon : MonoBehaviour {
   [Header("Settings")]
   public LayerMask buildingLayer;
-  public int directDamage;
-  public int sputteringDamege;
-  public int sputteringRadius;
+  public float directDamage;
+  public float sputteringDamege;
+  public float sputteringRadius;
   
 
   // Start is called before the first frame update
